# clock-bound-c: use array instead of vec for performance