@siddontang 2021-12-15
yes, `array`. sometimes I use `slice` by myself. I have changed it.