@jacoblwisniewski 2021-12-23
Tested it locally myself and confirmed it's working. Thank you for the optimization!